增加管理人员发送记录查看，修改整形BUG,修改不带通讯录的角色

+++ Rtdl.Basic.Data/Sms/_SmsChannel.cs
@@ -28,7 +28,7 @@ namespace Rtdl.Sms.Data
                     {
                         smsChannel e = new smsChannel
                         {
-                            ID = Convert.ToInt16(r["id"]),
+                            ID = Convert.ToInt32(r["id"]),
                             ChannelName = r["ChannelName"].ToString(),
                             MchUName = r["MchUName"].ToString(),
                             MchUPass = r["MchUPass"].ToString(),
